Fix for  python package dependencies, concurrent operations , container build issues

+++ hpedockerplugin/hpe_storage_api.py
@@ -43,6 +43,7 @@ import time
 DEFAULT_SIZE = 100
 DEFAULT_PROV = "thin"
 DEFAULT_FLASH_CACHE = None
+DEFAULT_MOUNT_VOLUME = "True"
 
 LOG = logging.getLogger(__name__)
 
@@ -132,14 +133,30 @@ class VolumePlugin(object):
         # Only 1 node in a multinode cluster can try to remove the volume.
         # Grab lock for volume name. If lock is inuse, just return with no
         # error.
-        self._lock_volume(volname, 'Remove')
+        # Expand lock code inline as function based lock causes
+        # unexpected behavior
+        try:
+            self._etcd.try_lock_volname(volname)
+        except Exception:
+            LOG.debug('volume: %(name)s is locked',
+                      {'name': volname})
+            response = json.dumps({u"Err": ''})
+            return response
 
         vol = self._etcd.get_vol_byname(volname)
         if vol is None:
             # Just log an error, but don't fail the docker rm command
             msg = (_LE('Volume remove name not found %s'), volname)
             LOG.error(msg)
-            self._unlock_volume(volname)
+            # Expand lock code inline as function based lock causes
+            # unexpected behavior
+            try:
+                self._etcd.try_unlock_volname(volname)
+            except Exception as ex:
+                LOG.debug('volume: %(name)s Unlock Volume Failed',
+                          {'name': volname})
+                response = json.dumps({u"Err": six.text_type(ex)})
+                return response
             return json.dumps({u"Err": ''})
 
         try:
@@ -150,7 +167,15 @@ class VolumePlugin(object):
             msg = (_LE('Err: Failed to remove volume %s, error is %s'),
                    volname, six.text_type(ex))
             LOG.error(msg)
-            self._unlock_volume(volname)
+            # Expand lock code inline as function based lock causes
+            # unexpected behavior
+            try:
+                self._etcd.try_unlock_volname(volname)
+            except Exception as ex:
+                LOG.debug('volume: %(name)s Unlock Volume Failed',
+                          {'name': volname})
+                response = json.dumps({u"Err": six.text_type(ex)})
+                return response
             raise exception.HPEPluginRemoveException(reason=msg)
 
         try:
@@ -161,7 +186,15 @@ class VolumePlugin(object):
             LOG.warning(msg)
             pass
 
-        self._unlock_volume(volname)
+        # Expand lock code inline as function based lock causes
+        # unexpected behavior
+        try:
+            self._etcd.try_unlock_volname(volname)
+        except Exception as ex:
+            LOG.debug('volume: %(name)s Unlock Volume Failed',
+                      {'name': volname})
+            response = json.dumps({u"Err": six.text_type(ex)})
+            return response
         return json.dumps({u"Err": ''})
 
     @app.route("/VolumeDriver.Unmount", methods=["POST"])
@@ -186,6 +219,11 @@ class VolumePlugin(object):
             LOG.error(msg)
             raise exception.HPEPluginUMountException(reason=msg)
 
+        vol_mount = DEFAULT_MOUNT_VOLUME
+        if ('Opts' in contents and contents['Opts'] and
+                'mount-volume' in contents['Opts']):
+            vol_mount = str(contents['Opts']['mount-volume'])
+
         path_info = self._etcd.get_vol_path_info(volname)
         if path_info:
             path_name = path_info['path']
@@ -203,10 +241,13 @@ class VolumePlugin(object):
         connector_info = connector.get_connector_properties(
             root_helper, self._my_ip, multipath=self.use_multipath,
             enforce_multipath=self.enforce_multipath)
-        # unmount directory
-        fileutil.umount_dir(mount_dir)
-        # remove directory
-        fileutil.remove_dir(mount_dir)
+
+        # Determine if we need to unmount a previously mounted volume
+        if vol_mount is DEFAULT_MOUNT_VOLUME:
+            # unmount directory
+            fileutil.umount_dir(mount_dir)
+            # remove directory
+            fileutil.remove_dir(mount_dir)
 
         # We're deferring the execution of the disconnect_volume as it can take
         # substantial
@@ -244,24 +285,6 @@ class VolumePlugin(object):
         response = json.dumps({u"Err": ''})
         return response
 
-    def _lock_volume(self, volname, op):
-        try:
-            self._etcd.try_lock_volname(volname)
-        except Exception:
-            LOG.debug('volume: %(name)s Volume %(op)s in progress',
-                      {'name': volname}, {'op': op})
-            response = json.dumps({u"Err": ''})
-            return response
-
-    def _unlock_volume(self, volname):
-        try:
-            self._etcd.try_unlock_volname(volname)
-        except Exception as ex:
-            LOG.debug('volume: %(name)s Unlock Volume Failed',
-                      {'name': volname})
-            response = json.dumps({u"Err": six.text_type(ex)})
-            return response
-
     @app.route("/VolumeDriver.Create", methods=["POST"])
     def volumedriver_create(self, name, opts=None):
         """
@@ -284,7 +307,8 @@ class VolumePlugin(object):
         volname = contents['Name']
 
         # Verify valid Opts arguments.
-        valid_volume_create_opts = ['size', 'provisioning', 'flash-cache']
+        valid_volume_create_opts = ['mount-volume',
+                                    'size', 'provisioning', 'flash-cache']
         if ('Opts' in contents and contents['Opts']):
             for key in contents['Opts']:
                 if key not in valid_volume_create_opts:
@@ -315,7 +339,15 @@ class VolumePlugin(object):
 
         # Grab lock for volume name. If lock is inuse, just return with no
         # error
-        self._lock_volume(volname, 'Create')
+        # Expand lock code inline as function based lock causes
+        # unexpected behavior
+        try:
+            self._etcd.try_lock_volname(volname)
+        except Exception:
+            LOG.debug('volume: %(name)s is locked',
+                      {'name': volname})
+            response = json.dumps({u"Err": ''})
+            return response
 
         # NOTE: Since Docker passes user supplied names and not a unique
         # uuid, we can't allow duplicate volume names to exist.
@@ -324,7 +356,15 @@ class VolumePlugin(object):
         vol = self._etcd.get_vol_byname(volname)
         if vol is not None:
             # Release lock and return
-            self._unlock_volume(volname)
+            # Expand lock code inline as function based lock causes
+            # unexpected behavior
+            try:
+                self._etcd.try_unlock_volname(volname)
+            except Exception as ex:
+                LOG.debug('volume: %(name)s Unlock Volume Failed',
+                          {'name': volname})
+                response = json.dumps({u"Err": six.text_type(ex)})
+                return response
             return json.dumps({u"Err": ''})
 
         voluuid = str(uuid.uuid4())
@@ -338,7 +378,15 @@ class VolumePlugin(object):
             # Release lock and return
             # NOTE: if for some reason unlock fails, we'll lose this
             # create exception.
-            self._unlock_volume(volname)
+            # Expand lock code inline as function based lock causes
+            # unexpected behavior
+            try:
+                self._etcd.try_unlock_volname(volname)
+            except Exception as ex:
+                LOG.debug('volume: %(name)s Unlock Volume Failed',
+                          {'name': volname})
+                response = json.dumps({u"Err": six.text_type(ex)})
+                return response
             return json.dumps({u"Err": six.text_type(ex)})
 
         response = json.dumps({u"Err": ''})
@@ -355,7 +403,15 @@ class VolumePlugin(object):
             LOG.error(msg)
             response = json.dumps({u"Err": six.text_type(ex)})
 
-        self._unlock_volume(volname)
+        # Expand lock code inline as function based lock causes
+        # unexpected behavior
+        try:
+            self._etcd.try_unlock_volname(volname)
+        except Exception as ex:
+            LOG.debug('volume: %(name)s Unlock Volume Failed',
+                      {'name': volname})
+            response = json.dumps({u"Err": six.text_type(ex)})
+            return response
         return response
 
     @app.route("/VolumeDriver.Mount", methods=["POST"])
@@ -385,6 +441,11 @@ class VolumePlugin(object):
             msg = (_LE('Volume mount name not found %s'), volname)
             LOG.error(msg)
             raise exception.HPEPluginMountException(reason=msg)
+
+        vol_mount = DEFAULT_MOUNT_VOLUME
+        if ('Opts' in contents and contents['Opts'] and
+                'mount-volume' in contents['Opts']):
+            vol_mount = str(contents['Opts']['mount-volume'])
 
         # Get connector info from OS Brick
         # TODO: retrieve use_multipath and enforce_multipath from config file
@@ -431,31 +492,37 @@ class VolumePlugin(object):
                   {'name': volname, 'device': device_info['path'],
                    'realpath': path.path})
 
-        # mkdir for mounting the filesystem
-        mount_dir = fileutil.mkdir_for_mounting(device_info['path'])
-        LOG.debug('Directory: %(mount_dir)s, successfully created to mount: '
-                  '%(mount)s',
-                  {'mount_dir': mount_dir, 'mount': device_info['path']})
-
         # Create filesystem on the new device
         if fileutil.has_filesystem(path.path) is False:
             fileutil.create_filesystem(path.path)
             LOG.debug('filesystem successfully created on : %(path)s',
                       {'path': path.path})
 
-        # mount the directory
-        fileutil.mount_dir(path.path, mount_dir)
-        LOG.debug('Device: %(path) successfully mounted on %(mount)s',
-                  {'path': path.path, 'mount': mount_dir})
+        # Determine if we need to mount the volume
+        if vol_mount is DEFAULT_MOUNT_VOLUME:
+            # mkdir for mounting the filesystem
+            mount_dir = fileutil.mkdir_for_mounting(device_info['path'])
+            LOG.debug('Directory: %(mount_dir)s, '
+                      'successfully created to mount: '
+                      '%(mount)s',
+                      {'mount_dir': mount_dir, 'mount': device_info['path']})
 
-        # TODO: find out how to invoke mkfs so that it creates the filesystem
-        # without the lost+found directory
-        # KLUDGE!!!!!
-        lostfound = mount_dir + '/lost+found'
-        lfdir = FilePath(lostfound)
-        if lfdir.exists and fileutil.remove_dir(lostfound):
-            LOG.debug('Successfully removed : %(lost)s from mount: %(mount)s',
-                      {'lost': lostfound, 'mount': mount_dir})
+            # mount the directory
+            fileutil.mount_dir(path.path, mount_dir)
+            LOG.debug('Device: %(path) successfully mounted on %(mount)s',
+                      {'path': path.path, 'mount': mount_dir})
+
+            # TODO: find out how to invoke mkfs so that it creates the
+            # filesystem without the lost+found directory
+            # KLUDGE!!!!!
+            lostfound = mount_dir + '/lost+found'
+            lfdir = FilePath(lostfound)
+            if lfdir.exists and fileutil.remove_dir(lostfound):
+                LOG.debug('Successfully removed : '
+                          '%(lost)s from mount: %(mount)s',
+                          {'lost': lostfound, 'mount': mount_dir})
+        else:
+            mount_dir = ''
 
         path_info = {}
         path_info['name'] = volname
@@ -466,7 +533,9 @@ class VolumePlugin(object):
 
         self._etcd.update_vol(volid, 'path_info', json.dumps(path_info))
 
-        response = json.dumps({u"Err": '', u"Mountpoint": mount_dir})
+        response = json.dumps({u"Err": '', u"Name": volname,
+                               u"Mountpoint": mount_dir,
+                               u"Devicename": path.path})
         return response
 
     @app.route("/VolumeDriver.Path", methods=["POST"])
@@ -519,11 +588,16 @@ class VolumePlugin(object):
         path_info = self._etcd.get_vol_path_info(volname)
         if path_info is not None:
             mountdir = path_info['mount_dir']
+            devicename = path_info['path']
         else:
             mountdir = ''
+            devicename = ''
 
-        volume = {'Name': volname,
+        # use volinfo as volname could be partial match
+        volume = {'Name': volinfo['display_name'],
                   'Mountpoint': mountdir,
+                  'Devicename': devicename,
+                  'Size': volinfo['size'],
                   'Status': {}}
 
         response = json.dumps({u"Err": err, u"Volume": volume})
@@ -550,11 +624,16 @@ class VolumePlugin(object):
                 path_info = self._etcd.get_path_info_from_vol(volinfo.value)
                 if path_info is not None and 'mount_dir' in path_info:
                     mountdir = path_info['mount_dir']
+                    devicename = path_info['path']
                 else:
                     mountdir = ''
+                    devicename = ''
                 info = json.loads(volinfo.value)
                 volume = {'Name': info['display_name'],
-                          'Mountpoint': mountdir}
+                          'Devicename': devicename,
+                          'size': info['size'],
+                          'Mountpoint': mountdir,
+                          'Status': {}}
                 volumelist.append(volume)
 
         response = json.dumps({u"Err": '', u"Volumes": volumelist})
